+++ extensions/deepcast/src/chinese.ts
@@ -1,6 +1,6 @@
 import { LaunchProps } from "@raycast/api";
 import { translate } from "./utils";
 
-export default async function Chinese(props: LaunchProps) {
-  await translate("ZH", props.fallbackText);
+export default async function Chinese(props: LaunchProps<{ arguments?: Arguments.Chinese }>) {
+  await translate("ZH", props.arguments?.text ?? props.fallbackText);
 }