+++ extensions/deepcast/src/finnish.ts
@@ -1,6 +1,6 @@
 import { LaunchProps } from "@raycast/api";
 import { translate } from "./utils";
 
-export default async function Finnish(props: LaunchProps) {
-  await translate("FI", props.fallbackText);
+export default async function Finnish(props: LaunchProps<{ arguments?: Arguments.Finnish }>) {
+  await translate("FI", props.arguments?.text ?? props.fallbackText);
 }