+++ extensions/deepcast/src/portuguese.ts
@@ -1,6 +1,6 @@
 import { LaunchProps } from "@raycast/api";
 import { translate } from "./utils";
 
-export default async function Portuguese(props: LaunchProps) {
-  await translate("PT-PT", props.fallbackText);
+export default async function Portuguese(props: LaunchProps<{ arguments?: Arguments.Portuguese }>) {
+  await translate("PT-PT", props.arguments?.text ?? props.fallbackText);
 }