+++ extensions/deepcast/src/romanian.ts
@@ -1,6 +1,6 @@
 import { LaunchProps } from "@raycast/api";
 import { translate } from "./utils";
 
-export default async function Romanian(props: LaunchProps) {
-  await translate("RO", props.fallbackText);
+export default async function Romanian(props: LaunchProps<{ arguments?: Arguments.Romanian }>) {
+  await translate("RO", props.arguments?.text ?? props.fallbackText);
 }