+++ extensions/deepcast/src/spanish.ts
@@ -1,6 +1,6 @@
 import { LaunchProps } from "@raycast/api";
 import { translate } from "./utils";
 
-export default async function Spanish(props: LaunchProps) {
-  await translate("ES", props.fallbackText);
+export default async function Spanish(props: LaunchProps<{ arguments?: Arguments.Spanish }>) {
+  await translate("ES", props.arguments?.text ?? props.fallbackText);
 }