+++ extensions/deepcast/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Deepcast Changelog
 
-## [Default Target Language and Fallback] - 2023-10-30
+## [Arguments] - 2023-10-28
+- Added support for inputting text to translate as an argument
 
+## [Default Target Language and Fallback] - 2023-10-16
 - Added preferences to set the default target language
 - Support commands being a fallback command